Delete codecov (#8)
Codecov 0.1.x were yanked from rubygems, causing installation of cs_comments_service to fail.

The initial approach was porting #5, which was ported from the PR against cs_comments_service master: edx#323.
However it's still not working with an older ruby version, so we decided to delete this dependency, as we're not using it anyway. It shouldn't actually be installed, as it's in the test group, but since this is a backport that we'll use only temporarily (before upgrading the instance) it didn't make too much sense to dig into this.